fix the segmentation fault problem and the task log time problem

+++ matrix/src/util.h
@@ -91,7 +91,7 @@ extern void gen_dag_indegree(adjList&, inDegree&);
 extern void gen_dag_parents(adjList&, adjList&);
 
 /* get the current time of day in micro-second */
-extern double get_time_usec();
+extern long get_time_usec();
 
 /* get the current time of day in mili-second */
 extern double get_time_msec();
